+++ R/FilterDISR.R
@@ -7,6 +7,13 @@
 #'
 #' This filter supports partial scoring (see [Filter]).
 #'
+#' @references
+#' `r format_bib("kursa_2021")`
+#'
+#' For a benchmark of filter methods:
+#'
+#' `r format_bib("bommert_2020")`
+#'
 #' @template details_praznik
 #' @family Filter
 #' @template seealso_filter
@@ -23,9 +30,9 @@ FilterDISR = R6Class("FilterDISR",
 
     #' @description Create a FilterDISR object.
     initialize = function() {
-      param_set = ParamSet$new(list(
-        ParamInt$new("threads", lower = 0L, default = 0L, tags = "threads")
-      ))
+      param_set = ps(
+        threads = p_int(lower = 0L, default = 0L, tags = "threads")
+      )
       param_set$values = list(threads = 1L)
 
       super$initialize(