+++ tests/testthat/test_FilterSelectedFeatures.R
@@ -1,8 +1,8 @@
-test_that("FilterImportance", {
+test_that("FilterSelectedFeatures", {
   set.seed(42)
   task = mlr3::mlr_tasks$get("wine")
   learner = mlr3::mlr_learners$get("classif.rpart")
-  f = FilterImportance$new(learner = learner)
+  f = FilterSelectedFeatures$new(learner = learner)
   f$calculate(task)
   expect_filter(f, task = task)
 })